webapp: Add Gist-based incident source
This commit adds Gist-based current and past incident data source support.

The incident data in the Gist is published by the incident-publisher
workflow periodically executing the Node.js app of the same name in this
repository.

Note that the incident issues are not directly pulled from the clients
because GitHub imposes a very low hourly limit (60 requests per hour) for
unauthenticated API requests.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>